feat(MemorySource): Extracted loop stuff from the memory source
