+++ nes-sources/private/MemorySource.hpp
@@ -59,36 +59,13 @@ public:
     [[nodiscard]] std::ostream& toString(std::ostream& str) const override;
 
 private:
-    /// One or two CSV files. When FILE_PATH_2 is set, the source alternates between the two
-    /// after REPLAYS_PER_FILE full passes of the current one — used by the adaptive benchmark
-    /// to simulate a workload-distribution shift on a deterministic schedule.
-    std::vector<std::string> filePaths;
-    bool loop;
-    uint64_t replaysPerFile;
-    /// Suffix-match (case-insensitive) of the schema field name whose UINT64 value should receive
-    /// a per-replay offset so the event-time watermark advances across loop iterations. Empty
-    /// string disables the rewrite (default).
-    std::string monotonicTimestampField;
+    std::string filePath;
     Schema schema;
     ParserConfig parserConfig;
     size_t bufferSizeInBytes;
-    /// Cached at setup() so fillTupleBuffer doesn't recompute.
-    size_t tupleSizeBytes{0};
-    /// Byte offset of the monotonic-timestamp field within a row, or -1 if disabled / not found.
-    int monotonicTimestampOffsetBytes{-1};
-    /// Per-file tuple counts (sum of tuples across the file's pre-formatted buffers). Used to
-    /// advance the global timestamp offset by exactly one cycle after each completed pass.
-    std::vector<uint64_t> tuplesPerFile;
-    /// Value to add to the monotonic timestamp field of every emitted tuple. Grows by
-    /// tuplesPerFile[currentFileIdx] after each pass; 0 on the very first cycle.
-    uint64_t globalTimestampOffset{0};
     std::atomic<size_t> totalTuplesEmitted{0};
-    /// Outer vector: one entry per loaded file. Inner vector: pre-formatted row-layout buffers.
-    std::vector<std::vector<TupleBuffer>> preFormattedBuffers;
-    size_t currentFileIdx{0};
-    uint64_t currentReplayCount{0};
-    std::vector<TupleBuffer>::iterator currentBufferIter;
-    std::shared_ptr<AbstractBufferProvider> bufferProvider;
+    std::vector<TupleBuffer> preFormattedBuffers;
+    std::vector<TupleBuffer>::iterator preFormattedBuffersIter;
 };
 
 struct ConfigParametersCSVMemory
@@ -98,38 +75,8 @@ struct ConfigParametersCSVMemory
         std::nullopt,
         [](const std::unordered_map<std::string, std::string>& config) { return DescriptorConfig::tryGet(FILEPATH, config); }};
 
-    /// Optional second CSV. When set, the source alternates between the two files; switching
-    /// from one to the other after REPLAYS_PER_FILE full passes of the current file.
-    static inline const DescriptorConfig::ConfigParameter<std::string> FILEPATH_2{
-        "file_path_2",
-        std::string{},
-        [](const std::unordered_map<std::string, std::string>& config) { return DescriptorConfig::tryGet(FILEPATH_2, config); }};
-
-    /// How many full passes of the current file before advancing to the next one. Default 1 means
-    /// "as soon as a file is fully drained, switch" — equivalent to seamless concatenation.
-    static inline const DescriptorConfig::ConfigParameter<uint64_t> REPLAYS_PER_FILE{
-        "replays_per_file",
-        uint64_t{1},
-        [](const std::unordered_map<std::string, std::string>& config) { return DescriptorConfig::tryGet(REPLAYS_PER_FILE, config); }};
-
-    /// When non-empty, names a UINT64 schema field whose value is incremented on every replayed
-    /// emit by `cycles_completed * tuples_per_file`. This keeps a downstream event-time watermark
-    /// advancing across LOOP cycles instead of getting stuck at the maximum value of the file.
-    /// Suffix-match is used (case-insensitive), so `timestamp` matches `BID$TIMESTAMP`.
-    static inline const DescriptorConfig::ConfigParameter<std::string> MONOTONIC_TIMESTAMP_FIELD{
-        "monotonic_timestamp_field",
-        std::string{},
-        [](const std::unordered_map<std::string, std::string>& config)
-        { return DescriptorConfig::tryGet(MONOTONIC_TIMESTAMP_FIELD, config); }};
-
-    static inline const DescriptorConfig::ConfigParameter<bool> LOOP{
-        "loop",
-        false,
-        [](const std::unordered_map<std::string, std::string>& config) { return DescriptorConfig::tryGet(LOOP, config); }};
-
     static inline std::unordered_map<std::string, DescriptorConfig::ConfigParameterContainer> parameterMap
-        = DescriptorConfig::createConfigParameterContainerMap(
-            SourceDescriptor::parameterMap, FILEPATH, FILEPATH_2, REPLAYS_PER_FILE, MONOTONIC_TIMESTAMP_FIELD, LOOP);
+        = DescriptorConfig::createConfigParameterContainerMap(SourceDescriptor::parameterMap, FILEPATH);
 };
 
 }
